# edukate-backend/src/main/kotlin/io/github/sanyavertolet/edukate/backend/dtos/ProblemDto.kt
package io.github.sanyavertolet.edukate.backend.dtos

import io.github.sanyavertolet.edukate.backend.entities.Problem

data class ProblemDto(
    val key: String,
    val code: String,
    val bookSlug: String,
    val isHard: Boolean,
    val tags: List<String>,
    val text: String,
    val subtasks: List<Problem.Subtask>,
    val images: List<String>,
    val status: Problem.Status,
    val hasResult: Boolean,
)