# spec/openapi-edukate-backend.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:5810
security:
- cookieAuth: []
...
paths:
...
  /api/v1/problems:
    get:
      tags:
      - Problems
      summary: Get problem list
      description: Retrieves a paginated list of problem metadata
      operationId: getProblemList
      parameters:
      - name: page
        in: query
        description: Page number (zero-based)
        required: false
        schema:
          type: string
          minimum: 0
      - name: size
        in: query
        description: Number of problems per page
        required: false
        schema:
          type: string
          exclusiveMinimum: 0
          maximum: 100
          minimum: 1
      - name: bookSlug
        in: query
        required: false
        schema:
          type: string
      - name: prefix
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - SOLVED
          - SOLVING
          - FAILED
          - NOT_SOLVED
      - name: isHard
        in: query
        required: false
        schema:
          type: boolean
      - name: hasPictures
        in: query
        required: false
        schema:
          type: boolean
      - name: hasResult
        in: query
        required: false
        schema:
          type: boolean
      responses:
        "200":
          description: Successfully retrieved problem list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProblemMetadata"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProblemMetadata"
        "401":
          description: Authentication required when filtering by status (other than
            NOT_SOLVED)
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProblemMetadata"
      security: []
  /api/v1/problems/{bookSlug}/{code}:
    get:
      tags:
      - Problems
      summary: Get problem by key
      description: Retrieves a specific problem by its book slug and code
      operationId: getProblem
      parameters:
      - name: bookSlug
        in: path
        description: Book slug
        required: true
        schema:
          type: string
      - name: code
        in: path
        description: Problem code
        required: true
        schema:
          type: string
      responses:
        "200":
          description: Successfully retrieved problem
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDto"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDto"
        "404":
          description: Problem not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDto"
      security: []
...
  /api/v1/problems/count:
    get:
      tags:
      - Problems
      summary: Count problems
      description: Returns the total number of problems matching the filter
      operationId: count
      parameters:
      - name: bookSlug
        in: query
        required: false
        schema:
          type: string
      - name: prefix
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - SOLVED
          - SOLVING
          - FAILED
          - NOT_SOLVED
      - name: isHard
        in: query
        required: false
        schema:
          type: boolean
      - name: hasPictures
        in: query
        required: false
        schema:
          type: boolean
      - name: hasResult
        in: query
        required: false
        schema:
          type: boolean
      responses:
        "200":
          description: Successfully retrieved problem count
          content:
            application/json:
              schema:
                type: integer
                format: int64
        "401":
          description: Authentication required when filtering by status (other than
            NOT_SOLVED)
          content:
            application/json:
              schema:
                type: integer
                format: int64
      security: []
  /api/v1/problems/by-prefix:
    get:
      tags:
      - Problems
      summary: Get problem codes by prefix
      description: Retrieves a list of problem codes that match the given prefix
      operationId: getProblemCodesByPrefix
      parameters:
      - name: prefix
        in: query
        description: The prefix to match problem codes against
        required: true
        schema:
          type: string
          minLength: 1
      - name: limit
        in: query
        description: Maximum number of results to return
        required: false
        schema:
          type: integer
          format: int32
          default: 5
          exclusiveMinimum: 0
      responses:
        "200":
          description: Successfully retrieved problem codes
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
      security: []
...
components:
  schemas:
...
    ProblemMetadata:
      type: object
      properties:
        key:
          type: string
        code:
          type: string
        bookSlug:
          type: string
        isHard:
          type: boolean
        tags:
          type: array
          items:
            type: string
        status:
          type: string
          enum:
          - SOLVED
          - SOLVING
          - FAILED
          - NOT_SOLVED
      required:
      - bookSlug
      - code
      - isHard
      - key
      - status
      - tags
...
    ProblemDto:
      type: object
      properties:
        key:
          type: string
        code:
          type: string
        bookSlug:
          type: string
        isHard:
          type: boolean
        tags:
          type: array
          items:
            type: string
        text:
          type: string
        subtasks:
          type: array
          items:
            $ref: "#/components/schemas/Subtask"
        images:
          type: array
          items:
            type: string
        status:
          type: string
          enum:
          - SOLVED
          - SOLVING
          - FAILED
          - NOT_SOLVED
        hasResult:
          type: boolean
      required:
      - bookSlug
      - code
      - hasResult
      - images
      - isHard
      - key
      - status
      - subtasks
      - tags
      - text
    Subtask:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
      required:
      - id
      - text
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      description: JWT authentication cookie
      name: X-Auth
      in: cookie